+++ Assets/Plugins/iOS/DDNAUnityNotificationsPlugin.h
@@ -16,16 +16,10 @@
 
 #include <Foundation/Foundation.h>
 
-// Unity native interface
-void _registerForPushNotifications();
-void _unregisterForPushNotifications();
-
 @interface DDNAUnityNotificationsPlugin : NSObject
 
 + (instancetype)sharedPlugin;
 
-- (void)registerForPushNotifications;
-- (void)unregisterForPushNotifications;
 - (BOOL)applicationDidLaunchWithRemoteNotification;
 - (NSDictionary *)getRemoteNotification;